Move JavaScript out of custom widget templates
Leaving the JavaScript scripts in the templates would cause the document
rendered by Django to be littered with small scripts when a form uses
many custom widgets. Instead, the required JavaScript scripts have been
moved to separate files which apply the scripts to all of the custom
widgets on the rendered page based on data attributes. Thus, one can
simply use {{ form.media }} in their form templates to load all the
necessary files in one place.