Add phone number inputmask and remove SSNField clean
Create a field using an inputmask to format phone numbers. This
inputmask only accepts US formatted numbers.

The clean method from SSNField has been removed. It uses an inputmask to
add dashes to SSNs. Previously, the clean method was used to strip the
dashes before querying the database. However, it appears that SSNs are
being stored with dashes in the database.